@@ -45,7 +45,7 @@ public class ASTSubtractNode extends SimpleNode {
      jjtGetChild(0).interpret();
      jjtGetChild(1).interpret();
 
-     stack[--top] = new Integer(((Integer)stack[top]).intValue() -
+     stack[--top] = Integer.valueOf(((Integer)stack[top]).intValue() -
                                 ((Integer)stack[top + 1]).intValue());
   }
 
