@@ -36,7 +36,7 @@ class MyLexer extends CalcInputParserTokenManager {
 
    /**
     * We redefined the lexical error reporting function so that it beeps
-    * and displays a message thru the GUI.
+    * and displays a message through the GUI.
     */
    protected void LexicalError() {
      CalcGUI.Error("ERROR (click 0)");
