@@ -49,6 +49,10 @@ public class SimpleNode implements Node {
   public static Node jjtCreate(JavaParser p, int id) {
     return new SimpleNode(p, id);
   }
+  
+  public int getId() {
+	  return id;
+  }
 
   public void jjtOpen() {
     first = parser.getToken(1);	// new
@@ -89,7 +93,7 @@ public class SimpleNode implements Node {
   }
 
   /** Accept the visitor. **/
-  public Object acceptChildren(JavaParserVisitor visitor, Object data) {
+  public Object childrenAccept(JavaParserVisitor visitor, Object data) {
     if (children != null) {
       for (int i = 0; i < children.length; ++i) {
         children[i].jjtAccept(visitor, data);
