@@ -1,4 +1,4 @@
-/* Generated By:JavaCC: Do not edit this line. SimpleCharStream.java Version 5.0 */
+/* Generated By:JavaCC: Do not edit this line. SimpleCharStream.java Version 7.0 */
 /* JavaCCOptions:STATIC=true,SUPPORT_CLASS_VISIBILITY_PUBLIC=true */
 /**
  * An implementation of interface CharStream, where the stream is assumed to
@@ -20,10 +20,12 @@ public class SimpleCharStream
   static protected char[] buffer;
   static protected int maxNextCharInd = 0;
   static protected int inBuf = 0;
-  static protected int tabSize = 8;
+  static protected int tabSize = 1;
+  static protected boolean trackLineColumn = false;
 
-  static protected void setTabSize(int i) { tabSize = i; }
-  static protected int getTabSize(int i) { return tabSize; }
+  static public void setTabSize(int i) { tabSize = i; }
+  static public int getTabSize() { return tabSize; }
+
 
 
   static protected void ExpandBuff(boolean wrapAround)
@@ -346,4 +348,4 @@ public class SimpleCharStream
     buffer = null;
   }
 }
-/* JavaCC - OriginalChecksum=2be772234ef663705e3245220a5a5c6e (do not edit this line) */
+/* JavaCC - OriginalChecksum=7e880fd78e1684b57e1e4d99f7425277 (do not edit this line) */
