@@ -1,5 +1,5 @@
-/* Generated By:JavaCC: Do not edit this line. Token.java Version 5.0 */
-/* JavaCCOptions:TOKEN_EXTENDS=,KEEP_LINE_COL=null,SUPPORT_CLASS_VISIBILITY_PUBLIC=true */
+/* Generated By:JavaCC: Do not edit this line. Token.java Version 7.0 */
+/* JavaCCOptions:TOKEN_EXTENDS=,KEEP_LINE_COLUMN=false,SUPPORT_CLASS_VISIBILITY_PUBLIC=true */
 /**
  * Describes the input token stream.
  */
@@ -86,6 +86,7 @@ public class Token implements java.io.Serializable {
   /**
    * Returns the image.
    */
+  @Override
   public String toString()
   {
     return image;
@@ -117,4 +118,4 @@ public class Token implements java.io.Serializable {
   }
 
 }
-/* JavaCC - OriginalChecksum=b68cbb4925e9712f2a198108314c1d2f (do not edit this line) */
+/* JavaCC - OriginalChecksum=c98f69671d3d2726981e73c5c72d4696 (do not edit this line) */
