@@ -1,8 +1,8 @@
+/* ParserTokenManager.java */
 /* Generated By:JavaCC: Do not edit this line. ParserTokenManager.java */
 
 /** Token Manager. */
-class ParserTokenManager implements ParserConstants
-{
+class ParserTokenManager implements ParserConstants {
 
   /** Debug output. */
   public static  java.io.PrintStream debugStream = System.out;
@@ -14,8 +14,7 @@ static private int jjStopAtPos(int pos, int kind)
    jjmatchedPos = pos;
    return pos + 1;
 }
-static private int jjMoveStringLiteralDfa0_0()
-{
+static private int jjMoveStringLiteralDfa0_0(){
    switch(curChar)
    {
       case 65:
@@ -30,66 +29,10 @@ static private int jjMoveStringLiteralDfa0_0()
          return 1;
    }
 }
-static final int[] jjnextStates = {
-};
 
 /** Token literal values. */
 public static final String[] jjstrLiteralImages = {
 "", "\101", "\102", "\103", "\104", };
-
-/** Lexer state names. */
-public static final String[] lexStateNames = {
-   "DEFAULT",
-};
-static protected SimpleCharStream input_stream;
-static private final int[] jjrounds = new int[0];
-static private final int[] jjstateSet = new int[0];
-static protected char curChar;
-/** Constructor. */
-public ParserTokenManager(SimpleCharStream stream){
-   if (input_stream != null)
-      throw new TokenMgrError("ERROR: Second call to constructor of static lexer. You must use ReInit() to initialize the static variables.", TokenMgrError.STATIC_LEXER_ERROR);
-   input_stream = stream;
-}
-
-/** Constructor. */
-public ParserTokenManager(SimpleCharStream stream, int lexState){
-   this(stream);
-   SwitchTo(lexState);
-}
-
-/** Reinitialise parser. */
-static public void ReInit(SimpleCharStream stream)
-{
-   jjmatchedPos = jjnewStateCnt = 0;
-   curLexState = defaultLexState;
-   input_stream = stream;
-   ReInitRounds();
-}
-static private void ReInitRounds()
-{
-   int i;
-   jjround = 0x80000001;
-   for (i = 0; i-- > 0;)
-      jjrounds[i] = 0x80000000;
-}
-
-/** Reinitialise parser. */
-static public void ReInit(SimpleCharStream stream, int lexState)
-{
-   ReInit(stream);
-   SwitchTo(lexState);
-}
-
-/** Switch to specified lex state. */
-static public void SwitchTo(int lexState)
-{
-   if (lexState >= 1 || lexState < 0)
-      throw new TokenMgrError("Error: Ignoring invalid lexical state : " + lexState + ". State unchanged.", TokenMgrError.INVALID_LEXICAL_STATE);
-   else
-      curLexState = lexState;
-}
-
 static protected Token jjFillToken()
 {
    final Token t;
@@ -113,6 +56,8 @@ static protected Token jjFillToken()
 
    return t;
 }
+static final int[] jjnextStates = {0
+};
 
 static int curLexState = 0;
 static int defaultLexState = 0;
@@ -134,9 +79,10 @@ public static Token getNextToken()
    {
       curChar = input_stream.BeginToken();
    }
-   catch(java.io.IOException e)
+   catch(Exception e)
    {
       jjmatchedKind = 0;
+      jjmatchedPos = -1;
       matchedToken = jjFillToken();
       return matchedToken;
    }
@@ -174,6 +120,31 @@ public static Token getNextToken()
   }
 }
 
+static void SkipLexicalActions(Token matchedToken)
+{
+   switch(jjmatchedKind)
+   {
+      default :
+         break;
+   }
+}
+static void MoreLexicalActions()
+{
+   jjimageLen += (lengthOfMatch = jjmatchedPos + 1);
+   switch(jjmatchedKind)
+   {
+      default :
+         break;
+   }
+}
+static void TokenLexicalActions(Token matchedToken)
+{
+   switch(jjmatchedKind)
+   {
+      default :
+         break;
+   }
+}
 static private void jjCheckNAdd(int state)
 {
    if (jjrounds[state] != jjround)
@@ -194,4 +165,89 @@ static private void jjCheckNAddTwoStates(int state1, int state2)
    jjCheckNAdd(state2);
 }
 
+    /** Constructor. */
+    public ParserTokenManager(SimpleCharStream stream){
+
+      if (input_stream != null)
+        throw new TokenMgrError("ERROR: Second call to constructor of static lexer. You must use ReInit() to initialize the static variables.", TokenMgrError.STATIC_LEXER_ERROR);
+
+    input_stream = stream;
+  }
+
+  /** Constructor. */
+  public ParserTokenManager (SimpleCharStream stream, int lexState){
+    ReInit(stream);
+    SwitchTo(lexState);
+  }
+
+  /** Reinitialise parser. */
+  
+  static public void ReInit(SimpleCharStream stream)
+  {
+
+
+    jjmatchedPos =
+    jjnewStateCnt =
+    0;
+    curLexState = defaultLexState;
+    input_stream = stream;
+    ReInitRounds();
+  }
+
+  static private void ReInitRounds()
+  {
+    int i;
+    jjround = 0x80000001;
+    for (i = 0; i-- > 0;)
+      jjrounds[i] = 0x80000000;
+  }
+
+  /** Reinitialise parser. */
+  static public void ReInit(SimpleCharStream stream, int lexState)
+  
+  {
+    ReInit(stream);
+    SwitchTo(lexState);
+  }
+
+  /** Switch to specified lex state. */
+  public static void SwitchTo(int lexState)
+  {
+    if (lexState >= 1 || lexState < 0)
+      throw new TokenMgrError("Error: Ignoring invalid lexical state : " + lexState + ". State unchanged.", TokenMgrError.INVALID_LEXICAL_STATE);
+    else
+      curLexState = lexState;
+  }
+
+
+/** Lexer state names. */
+public static final String[] lexStateNames = {
+   "DEFAULT",
+};
+
+/** Lex State array. */
+public static final int[] jjnewLexState = {
+   -1, -1, -1, -1, -1, 
+};
+static final long[] jjtoToken = {
+   0x1fL, 
+};
+static final long[] jjtoSkip = {
+   0x0L, 
+};
+static final long[] jjtoSpecial = {
+   0x0L, 
+};
+static final long[] jjtoMore = {
+   0x0L, 
+};
+    static protected SimpleCharStream  input_stream;
+
+    static private final int[] jjrounds = new int[0];
+    static private final int[] jjstateSet = new int[2 * 0];
+    private static final StringBuilder jjimage = new StringBuilder();
+    private static StringBuilder image = jjimage;
+    private static int jjimageLen;
+    private static int lengthOfMatch;
+    static protected int curChar;
 }
