@@ -1,5 +1,5 @@
-/* Generated By:JavaCC: Do not edit this line. Token.java Version 5.0 */
-/* JavaCCOptions:TOKEN_EXTENDS=,KEEP_LINE_COL=null,SUPPORT_CLASS_VISIBILITY_PUBLIC=false */
+/* Generated By:JavaCC: Do not edit this line. Token.java Version 7.0 */
+/* JavaCCOptions:TOKEN_EXTENDS=,KEEP_LINE_COLUMN=true,SUPPORT_CLASS_VISIBILITY_PUBLIC=false */
 /**
  * Describes the input token stream.
  */
@@ -95,6 +95,7 @@ class Token implements java.io.Serializable {
   /**
    * Returns the image.
    */
+  @Override
   public String toString()
   {
     return image;
@@ -126,4 +127,4 @@ class Token implements java.io.Serializable {
   }
 
 }
-/* JavaCC - OriginalChecksum=169c5b3c60b1aef218bf367bf1938a64 (do not edit this line) */
+/* JavaCC - OriginalChecksum=b61474106e11964da7fa2d3c40fc47a0 (do not edit this line) */
