@@ -1,4 +1,4 @@
-/* Generated By:JavaCC: Do not edit this line. SimpleCharStream.java Version 5.0 */
+/* Generated By:JavaCC: Do not edit this line. SimpleCharStream.java Version 7.0 */
 /* JavaCCOptions:STATIC=true,SUPPORT_CLASS_VISIBILITY_PUBLIC=true */
 /**
  * An implementation of interface CharStream, where the stream is assumed to
@@ -28,10 +28,12 @@ public class SimpleCharStream
   static protected char[] buffer;
   static protected int maxNextCharInd = 0;
   static protected int inBuf = 0;
-  static protected int tabSize = 8;
+  static protected int tabSize = 1;
+  static protected boolean trackLineColumn = true;
 
-  static protected void setTabSize(int i) { tabSize = i; }
-  static protected int getTabSize(int i) { return tabSize; }
+  static public void setTabSize(int i) { tabSize = i; }
+  static public int getTabSize() { return tabSize; }
+
 
 
   static protected void ExpandBuff(boolean wrapAround)
@@ -468,6 +470,7 @@ public class SimpleCharStream
     line = bufline[j];
     column = bufcolumn[j];
   }
-
+  static boolean getTrackLineColumn() { return trackLineColumn; }
+  static void setTrackLineColumn(boolean tlc) { trackLineColumn = tlc; }
 }
-/* JavaCC - OriginalChecksum=382b1af94ddaf8942012341b3b13600e (do not edit this line) */
+/* JavaCC - OriginalChecksum=052d2c8783a7a693ccde91d90feb1d3b (do not edit this line) */
