@@ -63,7 +63,7 @@ public class Token {
   }
 
   /**
-   * No-argument contructor
+   * No-argument constructor
    */
   public Token() {}
 
